requirements.txt
Pillow==6.2.0
pytesseract==0.3.4
pdf2image==1.13.1
